+++ csv/mod.ts
@@ -66,4 +66,5 @@
 export * from "./stringify.ts";
 export * from "./parse.ts";
 export * from "./stream.ts";
+export * from "./csv_parse_stream.ts";
 export * from "./csv_stringify_stream.ts";
